+++ src/components/progress/progress-questions-detail/progress.css
