+++ src/components/progress/submission-details/SubmissionDetails.vue
