+++ src/components/progress/submission-details/submission.css
