+++ src/components/progress/submission-details/submissionDetails.js
