+++ src/router/index.js
@@ -3,8 +3,8 @@ import CodeEditor from "../components/code-editor/CodeEditor.vue";
 import CodeEditorContainer from "../components/code-editor-container/CodeEditorContainer.vue"
 import QuestionsGroups from "../components/questions-group/QuestionsGroups.vue"
 import HomePage from"../components/home/HomePage.vue"
-import ProgressQuestions from"../components/progress/progressQuestionsDetail/ProgressQuestions.vue"
-import SubmissionDetails from"../components/progress/submissionDetails/SubmissionDetails.vue"
+import ProgressQuestions from"../components/progress/progress-questions-detail/ProgressQuestions.vue"
+import SubmissionDetails from"../components/progress/submission-details/SubmissionDetails.vue"
 import SessionCard from"../components/user-session/session-card/SessionCard.vue"
 import SessionActives from"../components/user-session/session-actives/SessionActives.vue"
 import LoginPage from "../components/auth/login/LoginPage.vue"
